Factor out routes and middleware from main server file
