+++ app/server/index.d.ts
@@ -1,0 +1,1 @@
+declare var GIT_COMMIT_HASH: string;
