+++ app/server/routes/index.ts
@@ -1,0 +1,4 @@
+export { default as core } from "./core";
+export { default as api } from "./api";
+export { default as profile } from "./profile";
+export { default as frontend } from "./frontend";
